requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cached-property==2.0.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
inflection==0.5.1
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlalchemy-orm==1.2.10
starlette==0.41.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.34.0
